refactor: move the remote loader to core-config
