+++ packages/core-config/src/loaders/index.ts
@@ -1,0 +1,4 @@
+import { FileLoader } from "./file-loader";
+import { RemoteLoader } from "./remote-loader";
+
+export { FileLoader, RemoteLoader };
